requirements.txt
requests
boto3
pydns
dnslib
tldextract